restarts db and creates applications stage table
