+++ alembic/env.py
@@ -8,6 +8,8 @@ from alembic import context
 from domain.entities.user import Base as UserBase
 from domain.entities.company import Base as CompanyBase
 from domain.entities.location import Base as LocationBase
+from domain.entities.application_stage import Base as ApplicationStageBase
+from domain.entities.job import Base as JobBase
 from settings import get_settings
 
 
@@ -27,11 +29,15 @@ if config.config_file_name is not None:
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-# target_metadata = UserBase.metadata
 
-# target_metadata = CompanyBase.metadata
 
-target_metadata = LocationBase.metadata
+target_metadata = [
+    UserBase.metadata,
+    CompanyBase.metadata,
+    LocationBase.metadata,
+    ApplicationStageBase.metadata,
+    #JobBase.metadata
+]
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired:
