--- alembic/versions/2f7367898ea2_alter_state_table_add_column_admin_code.py
@@ -1,32 +1,0 @@
-"""alter state table, add column admin code
-
-Revision ID: 2f7367898ea2
-Revises: fbbf84049a37
-Create Date: 2025-05-16 20:46:53.098683
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2f7367898ea2'
-down_revision: Union[str, None] = 'fbbf84049a37'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('states', sa.Column('admin_code', sa.Integer(), nullable=False))
-    op.create_index(op.f('ix_states_admin_code'), 'states', ['admin_code'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_states_admin_code'), table_name='states')
-    op.drop_column('states', 'admin_code')
-    # ### end Alembic commands ###
