--- alembic/versions/c4eac481b039_modifying_location_tables.py
@@ -1,59 +1,0 @@
-"""modifying location tables
-
-Revision ID: c4eac481b039
-Revises: f90cef16c664
-Create Date: 2025-05-10 12:43:50.586839
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c4eac481b039'
-down_revision: Union[str, None] = 'f90cef16c664'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('countries',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('common_name', sa.String(length=255), nullable=False),
-    sa.Column('code', sa.String(length=3), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_countries_code'), 'countries', ['code'], unique=True)
-    op.create_index(op.f('ix_countries_common_name'), 'countries', ['common_name'], unique=True)
-    op.create_table('states',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('country_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['country_id'], ['countries.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_states_name'), 'states', ['name'], unique=False)
-    op.create_table('cities',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('state_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['state_id'], ['states.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_cities_name'), 'cities', ['name'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_cities_name'), table_name='cities')
-    op.drop_table('cities')
-    op.drop_index(op.f('ix_states_name'), table_name='states')
-    op.drop_table('states')
-    op.drop_index(op.f('ix_countries_common_name'), table_name='countries')
-    op.drop_index(op.f('ix_countries_code'), table_name='countries')
-    op.drop_table('countries')
-    # ### end Alembic commands ###
