+++ alembic/versions/ddc2e4a28037_restart_db.py
@@ -1,0 +1,107 @@
+"""restart db
+
+Revision ID: ddc2e4a28037
+Revises: 
+Create Date: 2025-05-29 11:30:49.058040
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+import sqlalchemy_utils
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'ddc2e4a28037'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('email', sa.String(), nullable=False),
+    sa.Column('password', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
+    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
+    op.create_index(op.f('ix_users_name'), 'users', ['name'], unique=False)
+    op.create_table('companies',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('link', sqlalchemy_utils.types.url.URLType(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_companies_id'), 'companies', ['id'], unique=False)
+    op.create_index(op.f('ix_companies_name'), 'companies', ['name'], unique=True)
+    op.create_table('countries',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('common_name', sa.String(length=255), nullable=False),
+    sa.Column('code', sa.String(length=3), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_countries_code'), 'countries', ['code'], unique=True)
+    op.create_index(op.f('ix_countries_common_name'), 'countries', ['common_name'], unique=True)
+    op.create_table('states',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('code', sa.String(), nullable=False),
+    sa.Column('admin_code', sa.Integer(), nullable=False),
+    sa.Column('country_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['country_id'], ['countries.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_states_admin_code'), 'states', ['admin_code'], unique=False)
+    op.create_index(op.f('ix_states_code'), 'states', ['code'], unique=False)
+    op.create_index(op.f('ix_states_name'), 'states', ['name'], unique=False)
+    op.create_table('cities',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('state_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['state_id'], ['states.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_cities_name'), 'cities', ['name'], unique=False)
+    op.create_table('applications_stage',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('application_stage', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_applications_stage_application_stage'), 'applications_stage', ['application_stage'], unique=False)
+    op.create_index(op.f('ix_applications_stage_id'), 'applications_stage', ['id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_applications_stage_id'), table_name='applications_stage')
+    op.drop_index(op.f('ix_applications_stage_application_stage'), table_name='applications_stage')
+    op.drop_table('applications_stage')
+    op.drop_index(op.f('ix_cities_name'), table_name='cities')
+    op.drop_table('cities')
+    op.drop_index(op.f('ix_states_name'), table_name='states')
+    op.drop_index(op.f('ix_states_code'), table_name='states')
+    op.drop_index(op.f('ix_states_admin_code'), table_name='states')
+    op.drop_table('states')
+    op.drop_index(op.f('ix_countries_common_name'), table_name='countries')
+    op.drop_index(op.f('ix_countries_code'), table_name='countries')
+    op.drop_table('countries')
+    op.drop_index(op.f('ix_companies_name'), table_name='companies')
+    op.drop_index(op.f('ix_companies_id'), table_name='companies')
+    op.drop_table('companies')
+    op.drop_index(op.f('ix_users_name'), table_name='users')
+    op.drop_index(op.f('ix_users_id'), table_name='users')
+    op.drop_index(op.f('ix_users_email'), table_name='users')
+    op.drop_table('users')
+    # ### end Alembic commands ###
