--- alembic/versions/f90cef16c664_restarting_company.py
@@ -1,42 +1,0 @@
-"""restarting company
-
-Revision ID: f90cef16c664
-Revises: 240b60b76230
-Create Date: 2025-05-09 20:44:56.283272
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlalchemy_utils
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f90cef16c664'
-down_revision: Union[str, None] = '240b60b76230'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('companies',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('link', sqlalchemy_utils.types.url.URLType(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_companies_id'), 'companies', ['id'], unique=False)
-    op.create_index(op.f('ix_companies_name'), 'companies', ['name'], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_companies_name'), table_name='companies')
-    op.drop_index(op.f('ix_companies_id'), table_name='companies')
-    op.drop_table('companies')
-    # ### end Alembic commands ###
