--- alembic/versions/fbbf84049a37_alter_state_table_add_column_code.py
@@ -1,32 +1,0 @@
-"""alter state table, add column code
-
-Revision ID: fbbf84049a37
-Revises: c4eac481b039
-Create Date: 2025-05-16 19:12:38.837201
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'fbbf84049a37'
-down_revision: Union[str, None] = 'c4eac481b039'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('states', sa.Column('code', sa.String(), nullable=False))
-    op.create_index(op.f('ix_states_code'), 'states', ['code'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_states_code'), table_name='states')
-    op.drop_column('states', 'code')
-    # ### end Alembic commands ###
